// onefive-web/src/shared/enums/tags.ts
export enum Tags {
  FUNDAMENTALS = "FUNDAMENTALS",
  MARKET = "MARKET",
  MARKETING = "MARKETING",
  SALES = "SALES",
  SCALING_AND_GROWTH = "SCALING_AND_GROWTH",
  FUNDING_AND_INVESTMENT = "FUNDING_AND_INVESTMENT",
  LEGAL = "LEGAL",
  HUMAN_RESOURCES_AND_TEAM = "HUMAN_RESOURCES_AND_TEAM",
  PRODUCT = "PRODUCT",
  TECHNOLOGY = "TECHNOLOGY",
  CUSTOMER = "CUSTOMER",
}